requirements.txt
# Core dependencies
google-generativeai==0.5.0
requests==2.31.0
pillow==10.0.0

# For development
black==23.3.0
pylint==2.17.5
